% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble_for_rank}
\alias{bubble_for_rank}
\title{Make bubble plot for ranks}
\usage{
bubble_for_rank(
  data,
  x_name,
  y_name,
  rank_name,
  highlight_name,
  arrange_rows_by = NULL,
  x_order = NULL,
  x_labels = NULL,
  vline_at = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} in long format.}

\item{x_name}{Variable name in \code{data} to be plotted on x-axis (e.g.,
method names).}

\item{y_name}{Variable name in \code{data} to be plotted on y-axis (e.g.,
variable names).}

\item{rank_name}{Variable name in \code{data} to indicate rank of \code{x}.}

\item{highlight_name}{Variable name in \code{data} to indicate which bubbles
to highlight.}

\item{arrange_rows_by}{A level in \code{x_name}. Rows will be arranged by
rank for this method. Default is \code{NULL}, i.e., plot as-is.}

\item{x_order}{Ordering of columns. Default is \code{NULL}, i.e., based on
levels in \code{x}.}

\item{x_labels}{Labels to print on x-axis. Default is \code{NULL}, i.e., use
levels in \code{x}.}

\item{vline_at}{x-position(s) to draw vertical lines at, e.g., 1.5 to draw
between 1st and 2nd columns. Default is \code{NULL}, i.e., no line.}
}
\value{
Returns a \code{ggplot2} object to allow further modification.
}
\description{
Make bubble plot for ranks
}
\examples{
library(Figures)
data(df_rank)
head(df_rank)
bubble_for_rank(data = df_rank, y_name = "Variable", x_name = "method",
                rank_name = "mean_rank", highlight_name = "sig")
bubble_for_rank(data = df_rank,
                y_name = "Variable", x_name = "method", rank_name = "mean_rank",
                highlight_name = "sig",
                arrange_rows_by = "Linear XGBoost (AUC-based)")
bubble_for_rank(data = df_rank,
                y_name = "Variable", x_name = "method", rank_name = "mean_rank",
                highlight_name = "sig",
                arrange_rows_by = "Linear XGBoost (AUC-based)",
                x_order = c("Logistic", "Linear XGBoost (loss-based)",
                            "Linear XGBoost (AUC-based)", "Linear SVM"),
                vline_at = 1.5)
}
