% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\docType{data}
\name{df_rank}
\alias{df_rank}
\title{A \code{data.frame} to illustrate \code{\link{bubble_for_rank}}.}
\format{
An object of class \code{data.frame} with 84 rows and 5 columns.
}
\usage{
df_rank
}
\description{
A \code{data.frame} to illustrate \code{\link{bubble_for_rank}}.
}
\keyword{datasets}
